Refactor Python UniFFI e2e harness
